add windows version for transport/tcp.go
the _windows.go file will compile only when GOOS=windows and does not rely on CGO. It uses default go libraries and is proxy aware (from env)
